Feature: print a FASTA of marker genes for each bin
Added the ability to print a FASTA file (amino acid) that includes the sequences
of the genes that were identified as marker genes in each bin. Feature is
implemented in the qa subprogram as a new output format (number 10).

Each sequence is printed with an informative header that takes the form:

>$bin_id $contig_name geneId=$gene_num;start=$start;end=$end;strand=$strand marker=$marker;mstart=$mstart;mend=$mend

Where $ denotes a variable.

	Variables:
	- bin_id: FASTA bin the sequence comes from
	- contig_name: contig the sequence comes from
	- gene_num: gene number as assigned by Prodigal
	- start: gene start position on contig (nucleotide)
	- end: gene end position on contig (nucleotide)
	- strand: strand the gene is on, 1 or -1
	- marker: name of the marker that was identified as a match to the gene
	- mstart: start position on gene of alignment with marker (amino acid)
	- mend: end position on gene of alignment with marker (amino acid)